// frontend/src/app/api/models/file-reference-details.ts
/* tslint:disable */
/* eslint-disable */
export interface FileReferenceDetails {
  creationDate?: string;
  description?: null | string;
  downloadUrl?: null | string;
  fileName?: null | string;
  fileSizeInBytes?: null | number;
  id?: string;
  mimeType?: null | string;
  onPerson?: null | string;
  onSource?: null | string;
}